requirements.txt
arxiv==1.4.3
captum==0.6.0
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.6.0
et-xmlfile==1.1.0
feedparser==6.0.10
filelock==3.8.2
fonttools==4.38.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gitdb==4.0.10
gitpython==3.1.29
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.2.0
iopath==0.1.10
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==21.3
pandas==1.3.5
pathtools==0.1.2
pillow==9.3.0
portalocker==2.7.0
promise==2.3
protobuf==4.21.12
psutil==5.9.4
pyparsing==3.0.9
pytorch-model-summary==0.1.2
pytz==2022.6
pyyaml==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.12.1
setproctitle==1.3.2
sgmllib3k==1.0.0
shortuuid==1.0.11
smmap==5.0.0
tabulate==0.9.0
termcolor==2.2.0
threadpoolctl==3.1.0
timm==0.6.12
tokenizers==0.13.2
torchsummary==1.5.1
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.13
wandb==0.13.7
wcwidth==0.2.5
yacs==0.1.8
zipp==3.11.0